[move]: Image 엔티티 위치 변경 image -> image.entity 위치 변경
